pyproject.toml
[project]
name = "elfactory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nicolo.calcagno", email = "nicolo.calcagno@aitho.it" }
]
requires-python = ">=3.12"
dependencies = [
    "datapizza-ai>=0.0.9",
    "datapizza-ai-clients-openai>=0.0.9",
    "datapizza-ai-tools-duckduckgo>=0.0.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.151.0",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "ruff>=0.14.8",
]
